server/requirements.txt
flask
flask_discord
flask_cors
python-dotenv